+++ CHANGELOG.md
@@ -16,12 +16,16 @@ All user visible changes to this project will be documented in this file. This p
 - Renamed `AudioTrackConstraints` to `DeviceAudioTrackConstraints` and `MediaStreamSettings.audio()` to `MediaStreamSettings.device_audio()`. ([#225], [#224])
 - Made `RoomHandle`'s `mute_audio()`, `unmute_audio()`, `disable_audio()`, `enable_audio()`, `disable_remote_audio()`, `enable_remote_audio()` to accept optional `MediaSourceKind`. ([#225], [#224])
 - Dropped support for `i686-linux-android` target. ([#243])
+- Reworked `RoomCloseReason`: ([#244])
+    - Changed `RoomCloseReason.reason()` to return `RoomCloseKind` instead of `String`.
+    - Removed `RoomCloseReason.is_err()`.
 
 ### Added
 
 - `DisplayAudioTrackConstraints` and `MediaStreamSettings.display_audio()` to capture system audio (only works in Chrome and Windows desktop). ([#225], [#224])
 - `ConnectionHandle.get_state()` and `ConnectionHandle.on_state_change()` for tracking connection with remote member in P2P mode. ([#229])
 - `Jason.network_changed()` for instant reconnection of all signalling and ICE transports. ([#236])
+- `RoomCloseKind` enumeration to reflect why `Room` was closed. ([#244])
 
 ### Fixed
 
@@ -41,6 +45,7 @@ All user visible changes to this project will be documented in this file. This p
 [#238]: https://github.com/instrumentisto/medea-jason/pull/238
 [#242]: https://github.com/instrumentisto/medea-jason/pull/242
 [#243]: https://github.com/instrumentisto/medea-jason/pull/243
+[#244]: https://github.com/instrumentisto/medea-jason/pull/244
 [095368cb]: https://github.com/instrumentisto/medea-jason/commit/095368cb2e5dd6faca44882955b7cab6b19d8ad2
 [`cargo-ndk` 4.0 version]: https://github.com/bbqsrc/cargo-ndk/blob/v4.0.0/CHANGELOG.md#v400---2025-07-30